+++ Runtime/MockHMDBuildSettings.cs
@@ -33,9 +33,53 @@ namespace Unity.XR.MockHMD
         };
 
         /// <summary>
+        /// Foveation Gaze Simulation Mode.
+        /// </summary>
+        public enum FoveationGazeSimulationMode
+        {
+            /// <summary>
+            /// Gaze at the center of the view
+            /// </summary>
+            Disabled = 0,
+            
+            /// <summary>
+            /// Gaze at the center of the view
+            /// </summary>
+            StaticCenter,
+
+            /// <summary>
+            /// Gaze look left and right (both eyes use same direction)
+            /// </summary>
+            MovementLeftRight,
+
+            /// <summary>
+            /// Gaze with independent movement for both eyes (Left eyemove up/down and right eye moves left/right)
+            /// </summary>
+            MovementEyeIndependent,
+        };
+
+        /// <summary>
         /// Stereo rendering mode.
         /// </summary>
         public RenderMode renderMode;
+
+        [Header("Foveated Rendering")]
+        /// <summary>
+        /// Enable foveated rendering mode.
+        /// </summary>
+#if !UNITY_2022_3 && !UNITY_2023_3_OR_NEWER
+        [System.NonSerialized]
+#endif
+        public bool foveationEnabled = false;
+
+        /// <summary>
+        /// Foveation Gaze simulation mode to use
+        /// </summary>
+        [Tooltip("Select how the gaze directions should be generated.  \nNote: Some GPUs/API don't support having different data for both eyes when rendering using SinglePass, having different patterns in both eyes won't work for them.")]
+#if !UNITY_2022_3 && !UNITY_2023_3_OR_NEWER
+        [System.NonSerialized]
+#endif
+        public FoveationGazeSimulationMode gazeSimulationMode = FoveationGazeSimulationMode.Disabled;
 
         /// <summary>
         /// Runtime access to build settings.