+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0-preview.2] - 2024-04-25
+ * Added foveated rendering support for Metal.
+
+## [1.4.0-preview.1] - 2023-04-04
+ * Add support for Foveated Rendering API
+
 ## [1.3.1-preview.1] - 2022-12-21
  * Mac arm64 support
  * Added missing API documentation